Optimization, cleanup, commenting
The drawBitmap_i2c and drawChar_i2c now send each read byte to a new
routine "drawByte" which draws 8 pixels at once. This requires the
background color to be passed as argument.
The old drawBitmap_i2c that uses drawPixel is also available as an
override, missing the background color argument.  It is slower by 45% in
a 96x96 bitmap with approx. half its pixels blank. It would be slower
with more filled bitmaps.

#define USE_I2C_FONT has the following functionality:
- controls now the behaviour of the  virtual write(uint8_t) function. If
>0 the drawChar_i2c is used, else the stock drawChar.
- The non used function (drawChar or drawChar_i2c) does nothing, freeing
up a little bit the sketch size).
- Also if  USE_I2C_FONT>0 the  glcdfont loading and the setTextSize
function are omitted

Commented the new functions

Renamed drawBitmapInSketch to drawBitmapRAM